exit status equal to 2 whe  syntax error.

+++ tokenizer/tokenizer.c
@@ -151,7 +151,7 @@ t_token	*tester_token(char *line)
 	token_stream = tokenizer_t_tokens(tokens_strings, tokens_array->full_idx);
 
 	// Liberar cada string individualmente
-	for (int j = 0; j < tokens_array->full_idx; j++)
+	for (int j = 0; (size_t) j < tokens_array->full_idx; j++)
 		free(tokens_strings[j]);
 
 	free(tokens_array);
